fix: 냉장고 바구니 서비스와 재료 서비스 간 순환 참조 오류 수정
